reorganize react component host component
