+++ packages/Angular/Generic/skip-chat/src/lib/react-component-host/component-factory-result.ts
@@ -1,0 +1,9 @@
+
+/**
+ * Interface for component factory function that's generated by Skip
+ */
+export interface ComponentFactoryResult {
+  component: any; // React component
+  print?: () => void;
+  refresh?: (data?: any) => void;
+}
