+++ packages/Angular/Generic/skip-chat/src/lib/react-component-host/react-component-host.ts
@@ -1,325 +1,12 @@
 import { SimpleChanges, EventEmitter, ElementRef, Injectable } from '@angular/core';
-import { SkipComponentCallbacks, SkipComponentRootSpec, SkipComponentStyles, SkipComponentUtilities, SkipComponentChildSpec } from '@memberjunction/skip-types';
-import { LogError } from '@memberjunction/core';
-import { BaseSingleton } from '@memberjunction/global';
-
-/**
- * CDN URLs for external dependencies
- * These can be configured via environment variables in the future
- */
-const CDN_URLS = {
-  // Core React dependencies
-  BABEL_STANDALONE: 'https://unpkg.com/@babel/standalone@7/babel.min.js',
-  REACT: 'https://unpkg.com/react@18/umd/react.development.js',
-  REACT_DOM: 'https://unpkg.com/react-dom@18/umd/react-dom.development.js',
-  
-  // Ant Design dependencies
-  DAYJS: 'https://unpkg.com/dayjs@1.11.10/dayjs.min.js',
-  
-  // UI Libraries - Using UMD builds that work with global React
-  ANTD_JS: 'https://unpkg.com/antd@5.12.8/dist/antd.js',
-  ANTD_CSS: 'https://unpkg.com/antd@5.12.8/dist/reset.css',
-  REACT_BOOTSTRAP_JS: 'https://unpkg.com/react-bootstrap@2.9.1/dist/react-bootstrap.js',
-  BOOTSTRAP_CSS: 'https://cdn.jsdelivr.net/npm/bootstrap@5.3.2/dist/css/bootstrap.min.css',
-  
-  // Data Visualization
-  D3_JS: 'https://cdnjs.cloudflare.com/ajax/libs/d3/7.8.5/d3.min.js',
-  CHART_JS: 'https://cdnjs.cloudflare.com/ajax/libs/Chart.js/4.4.0/chart.umd.js',
-  
-  // Utilities
-  LODASH_JS: 'https://cdnjs.cloudflare.com/ajax/libs/lodash.js/4.17.21/lodash.min.js'
-};
-
-/**
- * Component metadata for component registry system
- */
-export interface ComponentMetadata {
-  /** List of child component names required by this component */
-  requiredChildComponents: string[];
-  /** Context for component resolution (e.g., 'CRM', 'Finance', 'Standard') */
-  componentContext: string;
-  /** Version of the component specification */
-  version: string;
-}
-
-/**
- * Component registration entry with metadata
- */
-interface ComponentRegistryEntry {
-  component: any;
-  metadata?: {
-    context?: string;
-    version?: string;
-    description?: string;
-  };
-}
-
-/**
- * Global component registry service for managing reusable React components
- * Extends BaseSingleton to ensure a truly global singleton instance across
- * the entire application, even if this code is loaded multiple times.
- */
-export class GlobalComponentRegistry extends BaseSingleton<GlobalComponentRegistry> {
-  private components = new Map<string, ComponentRegistryEntry>();
-  
-  protected constructor() {
-    super(); // Call parent constructor to register in global store
-  }
-  
-  /**
-   * Get the singleton instance
-   */
-  public static get Instance(): GlobalComponentRegistry {
-    return super.getInstance<GlobalComponentRegistry>();
-  }
-  
-  /**
-   * Register a component with a simple key
-   */
-  public register(key: string, component: any): void {
-    this.components.set(key, { component });
-  }
-  
-  /**
-   * Get a component by key
-   */
-  public get(key: string): any {
-    const entry = this.components.get(key);
-    return entry?.component;
-  }
-  
-  /**
-   * Register a component with metadata for versioning and context
-   */
-  public registerWithMetadata(
-    name: string, 
-    context: string, 
-    version: string, 
-    component: any,
-    description?: string
-  ): void {
-    const key = this.createKey(name, context, version);
-    this.components.set(key, {
-      component,
-      metadata: { context, version, description }
-    });
-    
-    // Also register without version for backwards compatibility
-    const contextKey = `${name}_${context}`;
-    if (!this.components.has(contextKey)) {
-      this.register(contextKey, component);
-    }
-  }
-  
-  /**
-   * Create a standardized key from component metadata
-   */
-  private createKey(name: string, context: string, version: string): string {
-    return `${name}_${context}_${version}`;
-  }
-  
-  /**
-   * Get all registered component keys (useful for debugging)
-   */
-  public getRegisteredKeys(): string[] {
-    return Array.from(this.components.keys());
-  }
-  
-  /**
-   * Clear all registered components
-   */
-  public clear(): void {
-    this.components.clear();
-  }
-  
-  /**
-   * Check if a component is registered
-   */
-  public has(key: string): boolean {
-    return this.components.has(key);
-  }
-  
-  /**
-   * Get component with fallback options
-   */
-  public getWithFallback(name: string, context: string, version: string): any {
-    // Try exact match first
-    let key = this.createKey(name, context, version);
-    if (this.has(key)) {
-      return this.get(key);
-    }
-    
-    // Try without version
-    key = `${name}_${context}`;
-    if (this.has(key)) {
-      return this.get(key);
-    }
-    
-    // Try global version
-    key = `${name}_Global`;
-    if (this.has(key)) {
-      return this.get(key);
-    }
-    
-    // Try just the name
-    if (this.has(name)) {
-      return this.get(name);
-    }
-    
-    return null;
-  }
-  
-  /**
-   * Remove a component from the registry
-   */
-  public remove(key: string): void {
-    this.components.delete(key);
-  }
-}
-
-/**
- * Configuration for a React component to be hosted in Angular
- */
-export interface ReactComponentConfig {
-  component: SkipComponentRootSpec 
-  
-  /** The HTML container element where the React component will be rendered */
-  container: HTMLElement;
-  
-  /** Data to pass to the component (e.g., entities, lists, etc.) */
-  data?: any;
-  
-  /** Callbacks for component lifecycle events */
-  callbacks?: SkipComponentCallbacks;
-  
-  /** Initial state for the component */
-  initialState?: any;
-  
-  /** Utilities to pass to the component */
-  utilities?: SkipComponentUtilities;
-  
-  /** Styles to pass to the component */
-  styles?: SkipComponentStyles;
-  
-  /** Component metadata for registry integration */
-  metadata?: ComponentMetadata;
-}
-
-/**
- * Interface for component factory function that's generated by Skip
- */
-export interface ComponentFactoryResult {
-  component: any; // React component
-  print?: () => void;
-  refresh?: (data?: any) => void;
-}
-
-/**
- * Default styles that match the Skip design system
- */
-const DEFAULT_STYLES: SkipComponentStyles = {
-  colors: {
-    // Primary colors - modern purple/blue gradient feel
-    primary: '#5B4FE9',
-    primaryHover: '#4940D4',
-    primaryLight: '#E8E6FF',
-    
-    // Secondary colors - sophisticated gray
-    secondary: '#64748B',
-    secondaryHover: '#475569',
-    
-    // Status colors
-    success: '#10B981',
-    successLight: '#D1FAE5',
-    warning: '#F59E0B',
-    warningLight: '#FEF3C7',
-    error: '#EF4444',
-    errorLight: '#FEE2E2',
-    info: '#3B82F6',
-    infoLight: '#DBEAFE',
-    
-    // Base colors
-    background: '#FFFFFF',
-    surface: '#F8FAFC',
-    surfaceHover: '#F1F5F9',
-    
-    // Text colors with better contrast
-    text: '#1E293B',
-    textSecondary: '#64748B',
-    textTertiary: '#94A3B8',
-    textInverse: '#FFFFFF',
-    
-    // Border colors
-    border: '#E2E8F0',
-    borderLight: '#F1F5F9',
-    borderFocus: '#5B4FE9',
-    
-    // Shadows (as color strings for easy use)
-    shadow: 'rgba(0, 0, 0, 0.05)',
-    shadowMedium: 'rgba(0, 0, 0, 0.1)',
-    shadowLarge: 'rgba(0, 0, 0, 0.15)',
-  },
-  spacing: {
-    xs: '4px',
-    sm: '8px',
-    md: '16px',
-    lg: '24px',
-    xl: '32px',
-    xxl: '48px',
-    xxxl: '64px',
-  },
-  typography: {
-    fontFamily: '-apple-system, BlinkMacSystemFont, "Inter", "Segoe UI", Roboto, sans-serif',
-    fontSize: {
-      xs: '11px',
-      sm: '12px',
-      md: '14px',
-      lg: '16px',
-      xl: '20px',
-      xxl: '24px',
-      xxxl: '32px',
-    },
-    fontWeight: {
-      light: '300',
-      regular: '400',
-      medium: '500',
-      semibold: '600',
-      bold: '700',
-    },
-    lineHeight: {
-      tight: '1.25',
-      normal: '1.5',
-      relaxed: '1.75',
-    },
-  },
-  borders: {
-    radius: {
-      sm: '6px',
-      md: '8px',
-      lg: '12px',
-      xl: '16px',
-      full: '9999px',
-    },
-    width: {
-      thin: '1px',
-      medium: '2px',
-      thick: '3px',
-    },
-  },
-  shadows: {
-    sm: '0 1px 2px 0 rgba(0, 0, 0, 0.05)',
-    md: '0 4px 6px -1px rgba(0, 0, 0, 0.1), 0 2px 4px -1px rgba(0, 0, 0, 0.06)',
-    lg: '0 10px 15px -3px rgba(0, 0, 0, 0.1), 0 4px 6px -2px rgba(0, 0, 0, 0.05)',
-    xl: '0 20px 25px -5px rgba(0, 0, 0, 0.1), 0 10px 10px -5px rgba(0, 0, 0, 0.04)',
-    inner: 'inset 0 2px 4px 0 rgba(0, 0, 0, 0.06)',
-  },
-  transitions: {
-    fast: '150ms ease-in-out',
-    normal: '250ms ease-in-out',
-    slow: '350ms ease-in-out',
-  },
-  overflow: 'auto'
-};
+import { SkipComponentCallbacks, SkipComponentRootSpec, SkipComponentStyles, SkipComponentChildSpec } from '@memberjunction/skip-types';
+import { LogError } from '@memberjunction/core'; 
+import { CDN_URLS } from './cdn-urls'; 
+import { GlobalComponentRegistry } from './component-registry';
+import { ComponentFactoryResult } from './component-factory-result';
+import { ReactComponentConfig } from './component-config';
+import { DEFAULT_STYLES } from './default-styles';
+ 
 
 /**
  * Host class for integrating Skip-generated React components into Angular applications.
