+++ packages/Angular/Generic/skip-chat/src/lib/skip-chat/skip-chat.component.html
@@ -147,11 +147,6 @@
                                                 (click)="showSharingDialog()"></span>
                                     </button>    
                                 }
-                                <!-- TEMPORARY TEST BUTTON -->
-                                <button kendoButton class="test-button" style="background-color: #ff0000; color: white;">
-                                    <span class="fa-solid fa-flask"
-                                            (click)="testButtonClick()"></span>
-                                </button>
                             </div>
                         </div>
                     }
