+++ packages/Angular/Generic/skip-chat/src/lib/skip-chat/skip-chat.component.ts
@@ -38,8 +38,7 @@ import { BaseAngularComponent } from '@memberjunction/ng-base-types';
 import { MJNotificationService } from '@memberjunction/ng-notifications';
 import { ResourcePermissionsComponent } from '@memberjunction/ng-resource-permissions';
 import { DrillDownInfo } from '../drill-down-info';
-import { SkipReactComponentHost, GlobalComponentRegistry } from '../dynamic-report/skip-react-component-host';
-import { WindowRef, DialogService, DialogRef } from '@progress/kendo-angular-dialog';
+import { DialogService, DialogRef } from '@progress/kendo-angular-dialog';
 
 @Component({
   selector: 'skip-chat',
@@ -2268,752 +2267,7 @@ export class SkipChatComponent extends BaseAngularComponent implements OnInit, A
         return [];
       }
     }
-  }
-  
-  /**
-   * Process a SkipComponentRootSpec and register all components recursively
-   */
-  private async processSkipComponentSpec(spec: any): Promise<{
-    success: boolean;
-    rootComponentName: string;
-    registeredComponents: string[];
-    errors: string[];
-  }> {
-    const errors: string[] = [];
-    const registeredComponents: string[] = [];
-    
-    try {
-      // Register root component (plain function, auto-wrapped)
-      const rootComponentName = spec.componentName;
-      const result = await SkipReactComponentHost.compileAndRegisterComponent(
-        rootComponentName,
-        spec.componentCode,
-        'Global', // Use Global context for all test components
-        'v1'
-      );
-      const success = result.success;
-      
-      if (success) {
-        registeredComponents.push(rootComponentName);
-        this.registeredTestComponents.set(rootComponentName, spec);
-      } else {
-        errors.push(`Failed to register root component: ${rootComponentName}`);
-      }
-      
-      // Recursively register child components (plain functions, auto-wrapped)
-      if (spec.childComponents && Array.isArray(spec.childComponents)) {
-        for (const child of spec.childComponents) {
-          await this.registerComponentHierarchy(child, registeredComponents, errors);
-        }
-      }
-      
-      return {
-        success: errors.length === 0,
-        rootComponentName,
-        registeredComponents,
-        errors
-      };
-    } catch (error: any) {
-      errors.push(`Error processing spec: ${error.message}`);
-      return {
-        success: false,
-        rootComponentName: spec.componentName || 'Unknown',
-        registeredComponents,
-        errors
-      };
-    }
-  }
-  
-  /**
-   * Recursively register child components
-   */
-  private async registerComponentHierarchy(
-    spec: any,
-    registeredComponents: string[],
-    errors: string[]
-  ): Promise<void> {
-    try {
-      if (spec.componentCode) {
-        // Child components are plain functions, auto-wrapped
-        const result = await SkipReactComponentHost.compileAndRegisterComponent(
-          spec.componentName,
-          spec.componentCode,
-          'Global', // Use Global context for all test components
-          'v1'
-        );
-        const success = result.success;
-        
-        if (success) {
-          registeredComponents.push(spec.componentName);
-        } else {
-          errors.push(`Failed to register component: ${spec.componentName}`);
-        }
-      }
-      
-      // Process nested children
-      const childArray = spec.childComponents || spec.components || [];
-      for (const child of childArray) {
-        await this.registerComponentHierarchy(child, registeredComponents, errors);
-      }
-    } catch (error: any) {
-      errors.push(`Error registering ${spec.componentName}: ${error.message}`);
-    }
-  }
-  
-  /**
-   * TEMPORARY TEST METHOD - Handle test button click
-   */
-  public async testButtonClick(): Promise<void> {
-    console.log('Test button clicked - Opening Component Tester');
-    
-    // If dialog already exists, just show it
-    if (this.componentTesterDialog) {
-      return;
-    }
-    
-    // Create the dialog container directly in the DOM
-    const dialogContainer = document.createElement('div');
-    dialogContainer.style.position = 'fixed';
-    dialogContainer.style.top = '50%';
-    dialogContainer.style.left = '50%';
-    dialogContainer.style.transform = 'translate(-50%, -50%)';
-    dialogContainer.style.width = '90vw';
-    dialogContainer.style.maxWidth = '1400px';
-    dialogContainer.style.height = '85vh';
-    dialogContainer.style.backgroundColor = 'white';
-    dialogContainer.style.boxShadow = '0 4px 20px rgba(0,0,0,0.3)';
-    dialogContainer.style.borderRadius = '8px';
-    dialogContainer.style.zIndex = '10000';
-    dialogContainer.style.display = 'flex';
-    dialogContainer.style.flexDirection = 'column';
-    
-    // Create header
-    const header = document.createElement('div');
-    header.style.padding = '16px 20px';
-    header.style.borderBottom = '1px solid #e0e0e0';
-    header.style.display = 'flex';
-    header.style.justifyContent = 'space-between';
-    header.style.alignItems = 'center';
-    header.style.backgroundColor = '#f5f5f5';
-    header.style.borderRadius = '8px 8px 0 0';
-    
-    const title = document.createElement('h3');
-    title.textContent = 'Skip Component Tester';
-    title.style.margin = '0';
-    title.style.fontSize = '18px';
-    title.style.fontWeight = '600';
-    
-    const closeButton = document.createElement('button');
-    closeButton.innerHTML = '×';
-    closeButton.style.fontSize = '28px';
-    closeButton.style.border = 'none';
-    closeButton.style.background = 'none';
-    closeButton.style.cursor = 'pointer';
-    closeButton.style.padding = '0';
-    closeButton.style.width = '32px';
-    closeButton.style.height = '32px';
-    closeButton.style.lineHeight = '1';
-    closeButton.onclick = () => {
-      // Clean up splitter event listeners
-      if (testerContent.refs.splitter && (testerContent.refs.splitter as any).cleanup) {
-        (testerContent.refs.splitter as any).cleanup();
-      }
-      
-      document.body.removeChild(dialogContainer);
-      document.body.removeChild(backdrop);
-      // Clear test components from registry
-      this.registeredTestComponents.forEach((spec, name) => {
-        if (!spec.prebuilt) {
-          // Only remove user-added components, not pre-built ones
-          GlobalComponentRegistry.Instance.remove(`${name}_Global_v1`);
-        }
-      });
-      this.registeredTestComponents.clear();
-      this.componentTesterDialog = null;
-    };
-    
-    header.appendChild(title);
-    header.appendChild(closeButton);
-    
-    // Create backdrop
-    const backdrop = document.createElement('div');
-    backdrop.style.position = 'fixed';
-    backdrop.style.top = '0';
-    backdrop.style.left = '0';
-    backdrop.style.right = '0';
-    backdrop.style.bottom = '0';
-    backdrop.style.backgroundColor = 'rgba(0,0,0,0.5)';
-    backdrop.style.zIndex = '9999';
-    
-    // Create the component tester content
-    const testerContent = this.createComponentTesterContent();
-    testerContent.element.style.flex = '1';
-    testerContent.element.style.overflow = 'hidden';
-    
-    dialogContainer.appendChild(header);
-    dialogContainer.appendChild(testerContent.element);
-    
-    // Add to DOM
-    document.body.appendChild(backdrop);
-    document.body.appendChild(dialogContainer);
-    
-    // Store reference for cleanup
-    this.componentTesterDialog = { close: () => closeButton.click() } as any;
-    
-    // Initialize code editor and other functionality
-    setTimeout(() => {
-      this.initializeComponentTester(testerContent);
-    }, 100);
-  }
-  
-  private createComponentTesterContent(): { element: HTMLElement; refs: any } {
-    const container = document.createElement('div');
-    container.style.height = '100%';
-    container.style.display = 'flex';
-    container.style.flexDirection = 'column';
-    
-    // Header with controls
-    const header = document.createElement('div');
-    header.style.padding = '16px';
-    header.style.borderBottom = '1px solid #e0e0e0';
-    header.style.display = 'flex';
-    header.style.gap = '16px';
-    header.style.alignItems = 'center';
-    
-    // Component selector dropdown
-    const selectorLabel = document.createElement('label');
-    selectorLabel.textContent = 'Component: ';
-    selectorLabel.style.fontWeight = 'bold';
-    
-    const componentSelector = document.createElement('select');
-    componentSelector.style.padding = '8px';
-    componentSelector.style.borderRadius = '4px';
-    componentSelector.style.border = '1px solid #ccc';
-    componentSelector.style.minWidth = '200px';
-    
-    // Add default option
-    const defaultOption = document.createElement('option');
-    defaultOption.value = '';
-    defaultOption.textContent = 'Select a component...';
-    componentSelector.appendChild(defaultOption);
-    
-    // Add new component button
-    const addNewBtn = document.createElement('button');
-    addNewBtn.textContent = 'Add New Component';
-    addNewBtn.className = 'k-button k-button-md k-rounded-md k-button-solid k-button-solid-primary';
-    addNewBtn.style.marginLeft = 'auto';
-    
-    // Parse & Register button
-    const parseBtn = document.createElement('button');
-    parseBtn.textContent = 'Parse & Register';
-    parseBtn.className = 'k-button k-button-md k-rounded-md k-button-solid k-button-solid-primary';
-    parseBtn.disabled = true;
-    
-    // Clear button
-    const clearBtn = document.createElement('button');
-    clearBtn.textContent = 'Clear All';
-    clearBtn.className = 'k-button k-button-md k-rounded-md k-button-solid k-button-solid-base';
-    
-    header.appendChild(selectorLabel);
-    header.appendChild(componentSelector);
-    header.appendChild(parseBtn);
-    header.appendChild(clearBtn);
-    header.appendChild(addNewBtn);
-    
-    // Main content area with split view
-    const content = document.createElement('div');
-    content.style.flex = '1';
-    content.style.display = 'flex';
-    content.style.overflow = 'hidden';
-    content.style.position = 'relative';
-    
-    // Get saved splitter position from local storage
-    const savedPosition = localStorage.getItem('skipComponentTesterSplitterPosition');
-    const leftWidth = savedPosition ? parseFloat(savedPosition) : 50;
-    const rightWidth = 100 - leftWidth;
-    
-    // Left panel - Code editor
-    const leftPanel = document.createElement('div');
-    leftPanel.style.width = `${leftWidth}%`;
-    leftPanel.style.display = 'flex';
-    leftPanel.style.flexDirection = 'column';
-    leftPanel.style.minWidth = '200px';
-    
-    const editorLabel = document.createElement('div');
-    editorLabel.style.padding = '8px 16px';
-    editorLabel.style.backgroundColor = '#f5f5f5';
-    editorLabel.style.borderBottom = '1px solid #e0e0e0';
-    editorLabel.innerHTML = '<strong>JSON Input (SkipComponentRootSpec)</strong>';
-    
-    const editorContainer = document.createElement('div');
-    editorContainer.style.flex = '1';
-    editorContainer.style.position = 'relative';
-    
-    leftPanel.appendChild(editorLabel);
-    leftPanel.appendChild(editorContainer);
-    
-    // Splitter handle
-    const splitter = document.createElement('div');
-    splitter.style.width = '6px';
-    splitter.style.backgroundColor = '#e0e0e0';
-    splitter.style.cursor = 'col-resize';
-    splitter.style.position = 'relative';
-    splitter.style.userSelect = 'none';
-    
-    // Add hover effect
-    splitter.onmouseenter = () => {
-      splitter.style.backgroundColor = '#c0c0c0';
-    };
-    splitter.onmouseleave = () => {
-      if (!isDragging) {
-        splitter.style.backgroundColor = '#e0e0e0';
-      }
-    };
-    
-    // Right panel - Component preview
-    const rightPanel = document.createElement('div');
-    rightPanel.style.width = `${rightWidth}%`;
-    rightPanel.style.display = 'flex';
-    rightPanel.style.flexDirection = 'column';
-    rightPanel.style.minWidth = '200px';
-    
-    const previewLabel = document.createElement('div');
-    previewLabel.style.padding = '8px 16px';
-    previewLabel.style.backgroundColor = '#f5f5f5';
-    previewLabel.style.borderBottom = '1px solid #e0e0e0';
-    previewLabel.innerHTML = '<strong>Component Preview</strong>';
-    
-    const previewContainer = document.createElement('div');
-    previewContainer.style.flex = '1';
-    previewContainer.style.overflow = 'auto';
-    previewContainer.style.padding = '16px';
-    
-    rightPanel.appendChild(previewLabel);
-    rightPanel.appendChild(previewContainer);
-    
-    // Add splitter drag functionality
-    let isDragging = false;
-    let startX = 0;
-    let startLeftWidth = 0;
-    
-    const handleMouseDown = (e: MouseEvent) => {
-      isDragging = true;
-      startX = e.pageX;
-      startLeftWidth = leftPanel.offsetWidth;
-      splitter.style.backgroundColor = '#a0a0a0';
-      
-      // Prevent text selection during drag
-      document.body.style.userSelect = 'none';
-      document.body.style.cursor = 'col-resize';
-      
-      e.preventDefault();
-    };
-    
-    const handleMouseMove = (e: MouseEvent) => {
-      if (!isDragging) return;
-      
-      const containerWidth = content.offsetWidth;
-      const deltaX = e.pageX - startX;
-      const newLeftWidth = startLeftWidth + deltaX;
-      const newLeftPercent = (newLeftWidth / containerWidth) * 100;
-      
-      // Limit the splitter position (20% to 80%)
-      if (newLeftPercent >= 20 && newLeftPercent <= 80) {
-        leftPanel.style.width = `${newLeftPercent}%`;
-        rightPanel.style.width = `${100 - newLeftPercent}%`;
-      }
-    };
-    
-    const handleMouseUp = () => {
-      if (!isDragging) return;
-      
-      isDragging = false;
-      splitter.style.backgroundColor = '#e0e0e0';
-      document.body.style.userSelect = '';
-      document.body.style.cursor = '';
-      
-      // Save position to local storage
-      const containerWidth = content.offsetWidth;
-      const leftPercent = (leftPanel.offsetWidth / containerWidth) * 100;
-      localStorage.setItem('skipComponentTesterSplitterPosition', leftPercent.toString());
-    };
-    
-    splitter.addEventListener('mousedown', handleMouseDown);
-    document.addEventListener('mousemove', handleMouseMove);
-    document.addEventListener('mouseup', handleMouseUp);
-    
-    // Clean up event listeners when dialog is closed
-    (splitter as any).cleanup = () => {
-      document.removeEventListener('mousemove', handleMouseMove);
-      document.removeEventListener('mouseup', handleMouseUp);
-    };
-    
-    // Error display area
-    const errorContainer = document.createElement('div');
-    errorContainer.style.display = 'none';
-    errorContainer.style.padding = '16px';
-    errorContainer.style.backgroundColor = '#fee';
-    errorContainer.style.color = '#c00';
-    errorContainer.style.borderTop = '1px solid #fcc';
-    errorContainer.style.maxHeight = '150px';
-    errorContainer.style.overflow = 'auto';
-    
-    content.appendChild(leftPanel);
-    content.appendChild(splitter);
-    content.appendChild(rightPanel);
-    
-    container.appendChild(header);
-    container.appendChild(content);
-    container.appendChild(errorContainer);
-    
-    return {
-      element: container,
-      refs: {
-        componentSelector,
-        addNewBtn,
-        parseBtn,
-        clearBtn,
-        editorContainer,
-        previewContainer,
-        errorContainer,
-        splitter
-      }
-    };
-  }
-  
-  private async initializeComponentTester(testerContent: { refs: any }): Promise<void> {
-    const { componentSelector, addNewBtn, parseBtn, clearBtn, editorContainer, previewContainer, errorContainer } = testerContent.refs;
-    
-    let currentHost: SkipReactComponentHost | null = null;
-    let codeEditorValue = '';
-    
-    // First, ensure React and other libraries are loaded by creating a temporary host
-    const tempContainer = document.createElement('div');
-    tempContainer.style.display = 'none';
-    document.body.appendChild(tempContainer);
-
-    // this is done to get React loaded
-    const tempSpec: SkipComponentRootSpec = {
-      componentCode: 'function TempComponent() {};',
-      componentName: 'TempComponent',
-      functionalRequirements: 'Load React and other libraries',
-      technicalDesign: 'Temporary component to load libraries',
-      componentType: 'report',
-      title: 'Temporary Component',
-      description: 'This component is used to load React and other libraries for testing purposes',
-      callbackStrategy: 'none',
-      childComponents: [],
-      stateStructure: {},
-      userExplanation: 'This component is used to load React and other libraries for testing purposes',
-      techExplanation: 'This component is used to load React and other libraries for testing purposes'
-    }
-    const tempHost = new SkipReactComponentHost({
-      component: tempSpec,
-      container: tempContainer,
-      data: {},
-      metadata: { requiredChildComponents: [], componentContext: 'Global', version: 'v1' }
-    });
-    
-    try {
-      await tempHost.initialize();
-      console.log('Libraries loaded via temporary host');
-    } catch (error) {
-      console.error('Error loading libraries:', error);
-    } finally {
-      tempHost.destroy();
-      document.body.removeChild(tempContainer);
-    }
-    
-    // Now register the pre-built test components
-    await this.registerPrebuiltComponents();
-    
-    // Add pre-built components to dropdown
-    const prebuiltComponents = [
-      { name: 'ActionBrowser', label: 'Action Browser (Pre-built)' },
-      { name: 'SearchBox', label: 'Search Box (Pre-built)' },
-      { name: 'CategoryChart', label: 'Category Chart (Pre-built)' }
-    ];
-    
-    for (const comp of prebuiltComponents) {
-      const option = document.createElement('option');
-      option.value = comp.name;
-      option.textContent = comp.label;
-      componentSelector.appendChild(option);
-      
-      // Ensure pre-built components are marked as such in our tracking map
-      if (!this.registeredTestComponents.has(comp.name)) {
-        this.registeredTestComponents.set(comp.name, { componentName: comp.name, prebuilt: true });
-      }
-    }
-    
-    // Create a textarea for now instead of the Angular component
-    const codeEditor = document.createElement('textarea');
-    codeEditor.style.width = '100%';
-    codeEditor.style.height = '100%';
-    codeEditor.style.fontFamily = 'monospace';
-    codeEditor.style.fontSize = '13px';
-    codeEditor.style.border = 'none';
-    codeEditor.style.outline = 'none';
-    codeEditor.style.resize = 'none';
-    codeEditor.style.padding = '16px';
-    codeEditor.style.backgroundColor = '#f5f5f5';
-    
-    // Listen for value changes
-    codeEditor.addEventListener('input', (event: any) => {
-      codeEditorValue = event.target.value;
-      parseBtn.disabled = !codeEditorValue.trim();
-    });
-    
-    editorContainer.appendChild(codeEditor);
-    
-    // Set initial value with example
-    const exampleSpec = {
-      componentName: "ExampleComponent",
-      componentType: "report",
-      title: "Example Component",
-      description: "An example component",
-      functionalRequirements: "Display hello world",
-      technicalDesign: "Simple React component",
-      callbackStrategy: "none",
-      stateStructure: {},
-      childComponents: [],
-      userExplanation: "Shows hello world",
-      techExplanation: "React createElement",
-      componentCode: `
-        function createComponent(React) {
-          const Component = () => {
-            return React.createElement('div', {
-              style: { padding: '20px', fontSize: '18px' }
-            }, 'Hello from Skip Component!');
-          };
-          return { component: Component };
-        }
-      `
-    };
-    
-    setTimeout(() => {
-      (codeEditor as any).value = JSON.stringify(exampleSpec, null, 2);
-      codeEditorValue = JSON.stringify(exampleSpec, null, 2);
-      parseBtn.disabled = false;
-    }, 200);
-    
-    // Handle component selection
-    componentSelector.addEventListener('change', async () => {
-      const selectedName = componentSelector.value;
-      if (!selectedName) {
-        if (currentHost) {
-          currentHost.destroy();
-          currentHost = null;
-        }
-        previewContainer.innerHTML = '<div style="color: #666; text-align: center; padding: 40px;">Select a component to preview</div>';
-        return;
-      }
-      
-      // Wait a bit to ensure components are registered
-      await new Promise(resolve => setTimeout(resolve, 100));
-      
-      // Debug: Check registry state
-      const registry = GlobalComponentRegistry.Instance;
-      console.log('Before rendering - Registry keys:', registry.getRegisteredKeys());
-      console.log('Looking for component:', selectedName);
-      
-      // Render the selected component
-      currentHost = await this.renderTestComponent(selectedName, previewContainer, currentHost);
-    });
-    
-    // Handle parse & register button
-    parseBtn.addEventListener('click', async () => {
-      errorContainer.style.display = 'none';
-      errorContainer.innerHTML = '';
-      
-      try {
-        const spec = JSON.parse(codeEditorValue);
-        
-        // Generate unique name if component already exists
-        let componentName = spec.componentName;
-        let counter = 1;
-        while (this.registeredTestComponents.has(componentName)) {
-          componentName = `${spec.componentName}_${counter}`;
-          counter++;
-        }
-        
-        const result = await this.processSkipComponentSpec(spec);
-        
-        if (result.success) {
-          // Add to dropdown
-          const option = document.createElement('option');
-          option.value = componentName;
-          option.textContent = `${componentName} (${result.registeredComponents.length} components)`;
-          componentSelector.appendChild(option);
-          componentSelector.value = componentName;
-          
-          // Trigger change event to render
-          componentSelector.dispatchEvent(new Event('change'));
-          
-          this.notificationService.CreateSimpleNotification(
-            `Successfully registered ${result.registeredComponents.length} components`,
-            'success',
-            3000
-          );
-        } else {
-          errorContainer.style.display = 'block';
-          errorContainer.innerHTML = '<strong>Errors:</strong><br>' + result.errors.join('<br>');
-        }
-      } catch (error: any) {
-        errorContainer.style.display = 'block';
-        errorContainer.innerHTML = '<strong>JSON Parse Error:</strong><br>' + error.message;
-      }
-    });
-    
-    // Handle clear button
-    clearBtn.addEventListener('click', () => {
-      // Clear only non-prebuilt components
-      this.registeredTestComponents.forEach((spec, name) => {
-        if (!spec.prebuilt) {
-          GlobalComponentRegistry.Instance.remove(`${name}_Global_v1`);
-          this.registeredTestComponents.delete(name);
-        }
-      });
-      
-      // Clear dropdown but keep pre-built options
-      componentSelector.innerHTML = '';
-      const defaultOption = document.createElement('option');
-      defaultOption.value = '';
-      defaultOption.textContent = 'Select a component...';
-      componentSelector.appendChild(defaultOption);
-      
-      // Re-add pre-built components
-      for (const comp of prebuiltComponents) {
-        const option = document.createElement('option');
-        option.value = comp.name;
-        option.textContent = comp.label;
-        componentSelector.appendChild(option);
-      }
-      
-      // Clear preview
-      if (currentHost) {
-        currentHost.destroy();
-        currentHost = null;
-      }
-      previewContainer.innerHTML = '<div style="color: #666; text-align: center; padding: 40px;">Select a component to preview</div>';
-      
-      // Clear editor
-      (codeEditor as any).value = '';
-      codeEditorValue = '';
-      parseBtn.disabled = true;
-    });
-    
-    // Handle add new button
-    addNewBtn.addEventListener('click', () => {
-      (codeEditor as any).value = JSON.stringify(exampleSpec, null, 2);
-      codeEditorValue = JSON.stringify(exampleSpec, null, 2);
-      parseBtn.disabled = false;
-    });
-    
-    // Initial state
-    previewContainer.innerHTML = '<div style="color: #666; text-align: center; padding: 40px;">Paste JSON and click "Parse & Register" to test a component</div>';
-  }
-  
-  private async renderTestComponent(
-    componentName: string,
-    container: HTMLElement,
-    currentHost: SkipReactComponentHost | null
-  ): Promise<SkipReactComponentHost | null> {
-    // Clean up existing host
-    if (currentHost) {
-      currentHost.destroy();
-    }
-    
-    container.innerHTML = '';
-    
-    try {
-      // Create wrapper component that uses the selected component
-      const wrapperCode = `
-        function createComponent(React, ReactDOM, useState, useEffect, useCallback, createStateUpdater, createStandardEventHandler, libraries) {
-          const WrapperComponent = ({ data, utilities, userState, callbacks, styles, components }) => {
-            const SelectedComponent = components['${componentName}'];
-            
-            if (!SelectedComponent) {
-              return React.createElement('div', {
-                style: { color: 'red', padding: '20px' }
-              }, 'Component "${componentName}" not found in registry');
-            }
-            
-            return React.createElement(SelectedComponent, {
-              data: data,
-              utilities: utilities,
-              userState: userState,
-              callbacks: callbacks,
-              styles: styles,
-              components: components
-            });
-          };
-          
-          return { component: WrapperComponent };
-        }
-      `;
-     
-      // The React host will handle all registration automatically
-      const host = new SkipReactComponentHost({
-        component: {
-          componentName: 'TestWrapper',
-          componentCode: wrapperCode,
-          componentType: 'other',
-          childComponents: [], // No child components for the wrapper
-          functionalRequirements: 'Test wrapper component',
-          technicalDesign: 'Wrapper for testing components',
-          description: 'Wrapper component for testing',
-          callbackStrategy: 'none',
-          stateStructure: {},
-          title: 'Test Wrapper',
-          userExplanation: 'This is a test wrapper',
-          techExplanation: 'Wraps components for testing'
-        } as SkipComponentRootSpec,
-        container: container,
-        data: {},
-        initialState: {},
-        utilities: {
-          md: {} as any,
-          rv: new TestRunView() as SimpleRunView,
-          rq: {} as any
-        },
-        metadata: {
-          requiredChildComponents: this.getRequiredChildComponents(componentName),
-          componentContext: 'Global', // Always use Global context for test components
-          version: 'v1'
-        },
-        callbacks: {
-          RefreshData: () => console.log('Refresh data requested'),
-          UpdateUserState: (state: any) => console.log('User state updated:', state),
-          OpenEntityRecord: (entityName: string, key: CompositeKey) => console.log('Open entity:', entityName, key),
-          NotifyEvent: (eventType: string, data: any) => console.log('Event:', eventType, data)
-        }
-      });
-      
-      await host.initialize();
-      console.log('Component rendered successfully');
-      
-      return host;
-    } catch (error) {
-      console.error('Error rendering component:', error);
-      container.innerHTML = `<div style="color: red; padding: 20px;">Error rendering component: ${error}</div>`;
-      return null;
-    }
-  }
+  }     
 }
 
-
-class TestRunView implements SimpleRunView {
-  private _rv: RunView;
-  constructor() {
-    this._rv = new RunView();
-  }
-  public runView(params: RunViewParams): Promise<RunViewResult> {
-    console.log('Running single view with params:', params);
-    return this._rv.RunView(params);
-  }
-  public runViews(params: RunViewParams[]): Promise<RunViewResult[]> {
-    console.log('Running multiple views with params:', params);
-    return this._rv.RunViews(params);
-  }
-}
+
